Shard py35 and py37 test cases
With more and more test cases the py35 and py37 test durations are
slowly approaching an hour which is currently our timeout. In order to
reduce the runtimes shard the test cases so we don't need to increase
the test timeouts even further.

This is done by listing the test modules in two shard files in
tests/unit. In order to prevent accidental skipping of test modules in
the future also validate that the shard files have all test modules
listed.

Change-Id: Icad29c203d6c75b764399d5a646862c1a4438c35