Docs: move overview section to reference
A user visiting the doc index and seeing that the first heading
is "Overview" might be forgiven for thinking that it would present
an overview of the Zuul project.  Instead, there is esoteric
reference information for people who are already quite involved.

Move these sections to the general reference section and remove
the "Overview" heading.

Change-Id: I2fc3672e5883e8fa89563a7282d1decef11e62ac