Docs: flatten directory structure
This drops the admins/ and users/ directories, so we are only left
with one level of organization (tutorials/howto/reference/discussion).
This way as we move things around, we may have slightly fewer redirects
to deal with.

This does not change the navigation structure or indexes.

Note, an alternative would be to drop tutorials/howto/etc and keep
admins/users, which was the original orginization, thus requiring even
fewer redirects.

Change-Id: I580088f5556330b76bc978be79baa0e9a2c552da