WIP: Fix mqtt log url reporting when report-build-page is active
Needs:
* docs
* test

Currently the mqtt reporter uses the report url as log_url. This is
fine as long as report-build-page is disabled. As soon as
report-build-page is enabled on a tenant it reports the url to the
result page of the build. As mqtt is meant to be consumed by machines
this breaks e.g. log post processing.

Fix this by reporting the real log url as log_url and add the field
web_url for use cases where really the human url is required.

Change-Id: I91ce93a7000ddd0d70ce504b70742262d8239a8f