Handle draft pull requests in canMerge
Github now supports draft pull requests which are blocked from merging
by github. Currently those can end up in a gate loop if all
requirements for the gate pipeline are fulfilled except for
un-drafting it.  Zuul needs to handle those in the canMerge check in
order to handle them correctly.

Change-Id: Ie9164ad5127d19ca3d75660b9718979da7cc344c