Update patch set 4
Patch Set 4:

Yes, the test failure on Xenial is related to this change. It's due to an old git version. I don't have an idea to work around that. The other way to fix this problem, is to configure the prune option via the git config. But I think that this should be fixed in Zuul.

Patch-set: 4