tests: throttle after all builds have reported
While profiling a test run, I have noticed a very high count of calls to
haveAllBuildsReported() and threading.Lock.acquire().

When not all builds have reported and there are events in the scheduler,
the while loop actually never waits. threading.Event().wait(0.1) returns
immediately.

"manually" wait when there are events in the scheduler.

Change-Id: Ic90709cd455b09eb750aa56fc306e3133f08a1de