Flesh out the glossary significantly
Zuul's complex design and novel underlying concepts require a
plethora of custom terminology to explain. This terminology is
carefully chosen in order to hopefully be identifiable and unique
within the CI/CD landscape, but is often not sufficiently
self-explanatory. We must try to incorporate as much into the
glossary as we can, in order to empower us to more clearly
communicate these concepts to newcomers.

Change-Id: If92045a01046572ad28631942baca6a87fdda2b6