ansible manager: only failed if last ansible failed
When validating several ansible versions, the validate() method returns
the output of the LAST `ansible --version` that has been executed.

The reason is the 'result' variable was used both for the stdout and to
keep track of a failure.

I have found that due to my 2.7 ansible being broken somehow and
validate was still happy since a later version was working all fine.

Add a test to ensure validate() fails properly in such a case.

Change-Id: Ib8bef996639e09f6355830c65f341d88f5cab8d8