web: show duration of each task in console
In the job console, show the duration of each task. It comes handy to
find potential slowdown in a task execution.

Each task has a start and end times, use momentjs to figure out the
delta and make it human friendly.

Note: durations have a threshold so that 2300 miliseconds show as
"in a few seconds". Seems good enough.

Thanks to Tristan Cacqueray for the guidance!

Change-Id: I955583713778911a8e50f08dc6d077594da4ae44