Update patch set 1
Patch Set 1: Code-Review-1

I like the idea, but I think it needs to be more precise.  The first job I picked just said "in a few seconds" over and over, and then "in an hour" at the bottom.  "An hour" might be okay for how long ago someone posted a news story, but it's not helpful in analyzing the run time of a computing process.  Using "HH:MM:SS", or "x hours, y minutes, z seconds", or "12h 34m 56s" is much more useful, and plenty readable.

I'm partial to the last option, which in most cases would look like:

* "in 38s"
* "in 34m 2s"

But another option might be to change the scale based on the length, like this:

* if duration < 60s:  "34 seconds"
* if 2h < duration > 60s:  "27 minutes"
* if duration > 2h:  "2 hours, 30 minutes"

Also, we could add some judgement in and highlight items which we consider long.

Patch-set: 1
Label: Code-Review=-1