Update patch set 17
Patch Set 17:

If we go back to the original proposal and store things in the builds table flat as you suggest, that makes the most sense to me from a data storage perspective. But then we still need to clean up places in the UI where we filter things. If instead we do Jim's suggestion of a prior_builds attribute in the result from the model - it feels like we'd get the best of both worlds... we wouldn't need to clean anything consume-side, but we wouldn't be double-storing things. Or?

Patch-set: 17
Label: Verified=0