Update patch set 17
Patch Set 17:

Yeah, I agree with Felix and Monty, I think the second table isn't great.  I'm sorry I suggested it and sent you down a bad path.

So let's expect to continue to have a single builds table in the db.  We could add in an extra field like "non_final" or "superceded" or something to make it easy to filter out previous builds, in case we want to store other kinds of non-final builds.  Or, we can keep things simple and just examine the result and look for RETRY since that's the only current case.

I still like the idea of keeping the web API easy to understand by having both a "builds" and "prior_builds" (or we could call it "non_final_builds" or similar) attribute, so that end-users don't have to account for the addition of the retry builds (and most won't be interested in them).  This would match the schema in the current version of the mqtt patch too.

Patch-set: 17
Label: Verified=0