Update patch set 17
Patch Set 17:

Thanks for the response and no problem about the "bad path" - I also just realized that after I did the actual change :D

I will do the following:
1. Restore this change to the old version
2. Add a field "superceded" to the builds table indicating that this build was either retried or reset (considering the follow-up change about storing gate resets).
3. Use the superceded field to filter out those builds in the web UI (to keep its state for now).

In a follow-up change we could use the superceded in combination with the build.result to list the builds in different sections on the buildset result page.
For the builds page we could hide the superceded builds by default and provide e.g. a checkbox next to the filter box saying "Show retried builds".

I think the last two points only make sense for retried builds as the gate resets are nearly transparent to the user.

Patch-set: 17
Label: Verified=0