Allow different branches for config projects
Currently we only allow the master branch for config projects. However
some projects need to use a different branch. In order to faciliotate
this use case allowe different branches for config projects as long as
there is exactly one branch. If there are multiple branches keep the
fall back to master.

This also prepares the use case of pinning a config repo to a specific
bramch via the tenant configuration.

Change-Id: Ic5040ce7ff71d857a5549dc61592e64cbdf04d6a