Store information about gate resets in MQTT and SQL reporter
When a build was canceled due to a gate reset, this build information is
lost. However, it would be helpful to get those information, especially
when analyzing build times. Therefore, this change stores those
information on the queue item and reports the builds in the SQL and MQTT
reporter.

Those "reset" builds are marked as non-final the same way like the
retried ones and thus, hidden from the UI by default.

Change-Id: Id3d715dbe0155172f66cecc19a54b3853c17a657