Update patch set 1
Patch Set 1:

If you had a single branch called "devel" and then later added a "master" branch, Zuul would immediately switch to the master branch.  Are we okay with that, or do we want to make it explicit (perhaps using the next patch) to opt in to a branch other than master in order to prevent this situation?

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0