Update patch set 1
Patch Set 1:

My initial intention was that a branch filter is generic and could be applied to untrusted projects as well. However I admit that this sacrifices usability on the config project use case. I'll switch the followup change to the explicit load-branch which will probably make this change obsolete.

Patch-set: 1