Allow more k8s actions in untrusted context
Allow the use of the k8s module in the untrusted execution context,
but check the 'src' attribute to make sure it isn't trying to load
a file from outside the executor work dir.  Otherwise, this should
be safe since it only operates on external resources.

Also, the add_host module can already be used with kubectl, but
it does not allow setting the namespace which may be necessary
if the credential being used isn't already scoped to the target
namespace.  Allow the ansible_kubectl_namespace option to be
used with add_host.

Change-Id: Id4f4dda2f7b743752553718aec69507bbc7a5b12