Scheduler: make autohold hold_list configurable
autohold makes it possible for a developer to help debug a failure by
offering access the failed live environment.
Sometimes it may happen that logs and live inspection are not enough to
discover a bug, and a step by step approach may be preferable.
A step by step requires to run only a small part of the build known to
work, then proceeding manually to analyze the next steps.
Currently in zuul this is not possible, as the node can be held only
when the build has failed, and if we configure the build to do a step
know to work, the node is freed at the end.
This change make the hold list of build results to allow autohold,
configurable, defaulting to the current harcoded value

Change-Id: Ifdf20bf6fe8808ee627c010e6fe69795530cea0f