Make most test cases work on MacOS
When developing on MacOS currently running test cases is only possible
via remote execution on a linux box. With latest master version of
gear it is possible to run zuul test cases also natively on MacOS with
minimal modification. Since bwrap is not available on mac add a small
fake bwrap script that can be linked into the PATH to make test cases
work that use ansible.

Change-Id: I02d7cc30b872c7406cf6641762e1f7b89e99479d