Reduce gearman logging in tests
Gearman produces a lot of logging in the tests which makes them large
and often hard to read. Since gear doesn't change often the logging it
produces could be reduced. Since there are many test cases that
produce a lot of logging this also reduces the risk of hitting the log
length limit of stestr when a test fails.

Change-Id: I095025a3072821f051c4b8a3fb8eb981c36992bb