Add handler only if it's not there already
The base test class currently adds a handler for every logger
specified in OS_LOG_DEFAULTS in every test case. Those don't get
cleaned so they accumulate with each test case and multiply logging
which leads to slow test cases and high memory consumption.

This has been hidden in ci so far because of a bug in the tox
configuration that sets the log config explicitly to an empty string
overriding the default logging config specified in the base test
class.

Fix this by adding the missing cleanup to remove the handler from
logging.

Change-Id: I1359acd7aef7e22505151f66915348bae1b1c740