Update patch set 1
Patch Set 1:

The goal of Zuul's configuration language is to have a straightforward language that any developer can read and understand easily.  We found that while JJB was mostly easy to follow, once we added significant templating and substitution, many people had difficulty understanding how the jobs ended up expanding out.

That's a risk here too, and I think it's worth considering whether the additional difficulty of defining a series of "one-liner" jobs:

- {job: {name: foo-linux-py27, parent: foo-linux, vars: {tox: py27}}}

is greater than the difficulty that users would face attempting to understand the matrix expansien series.  I see the need (indeed, I wrote a script in zuul-jobs to create jobs like this), but I'm not sure I'm convinced we can make something generalized and simple.

However, if we do think we need matrix jobs of some sort, then I would prefer to handle it in a way similar to this.  If the Zuul configuration language is deficient enough to need to be replaced or augmented with dhall, we should fix the Zuul configuration language -- dhall is not friendly or understandable for most users.

If we do proceed with something like this, we should try to keep it as simple as possible (it's okay if it doesn't handle every case; it's okay if it doesn't recurse).

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0