Update patch set 1
Patch Set 1:

I agree with Jim and Tristan. I think there is a risk in adding templating complexity - although if we do decide it's a direction we want to go in, I think this feels like a a better direction to go in than something more expansive or turing complete like jinja or dhall.

I'll say that I like that it's modeled after the tox.ini matrix, since that's actually now been around for a bit and seems to be something that users interact with successfully.

It's also nice that the implementation only took 38 lines.

Patch-set: 1
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0