configloader: validate playbook-less + secret jobs
If you create a job that has no playbooks, uses a secret but does
not make use of pass-to-parent, the secret is absolutely never used in
any circumstance.

This patch prevents Zuul users from getting potentially caught in this
issue by preventing them from merging a job configuration that has this.

Change-Id: I2d01bc204de6a1b6d2589678c8406cc780255b8f