WIP: store secret names in console json
The goal is to display in the web UI console log a list of secrets
available to each playbook.  This adds an extra variable passed
to ansible so that the callback can store that info.

This is the first non-scalar value (it's a list) passed to the
callback this way.  Rather than attempt to pass it as a command
line argument like the other zuul_execution vars, move all of them
into a yaml file which is then included via "-e @...".  This
makes it easier to pass structured data (it also means these
particular arguments no longer appear in the Ansible command line).

The resulting data is a list of variable names which correspond to
secrets.  It should not include any values, and variable names
themselves should not be secret.  I don't believe variable names
themselves can be jinja templates, so I don't think this is a path
to potentially expose a secret.  I also don't believe that any
expansion happens in the callback plugin, which should protect from
the same risk.

We may instead want this to be a list of Zuul Secret object names
rather than a list of final variable names.  Either is likely to
aid in debugging missing secret issues, but Zuul Secret object names
additionally answer the question "which of the multiple secrets
which contain these variables was used".  We could also hyperlink
to secret definitions in the web UI in the future.

Change-Id: I45425b9cdefcc4e2ba1cc132e4bc7fd8a017e4b0