ci: use bionic for functional tests
The functional tests seem to be currently running in Xenial which
is quite old and most likely most Zuul deployments are running on
Bionic at this point.

This updates the nodeset name to drop the 'temp' (as this nodeset seems
to be far from temporary) and uses Bionic instead of Xenial.

Change-Id: Ic2896e2949928daedf22dbac2a162039e70696f4