Dequeue changes via github checks API
The Github checks API allows us to define custom actions on a check run.
With that, we can define a custom "abort" action, that enables users to
directly abort a running change by clicking on a respective button in
Github's checks tab.

A click on this button will emit a Github webhook event that is now
handleded by Zuul to dequeue the respective change.

To uniquely identify a change represented by a check run, each check run
now comes with an external ID which is a combination of tenant,
pipeline, queue and PR number.

Change-Id: I574479cd06abd1ea4f3daa7850cf35ea6715a23d