Update patch set 17
Patch Set 17:

Tristan, I don't think it's too late for that -- if we need to switch, now would be better than later (or, at the very least, if we think we might need to switch later, we might implement it slightly differently).  Given that we've decided that the difference between zk and etcd for our current use is insignificant, do we think with this change in how we use ZK that we would see a difference?  It looks like etcd may use less memory.  Should we come up with some estimates for what we expect data size/traffic to reach, and run benchmarks on both?

Patch-set: 17