Update patch set 18
Patch Set 18: Code-Review+2

Discussion about zk evaluation continued a bit on IRC here: http://eavesdrop.openstack.org/irclogs/%23zuul/%23zuul.2020-02-21.log.html#t2020-02-21T15:28:10

Considering zuul and nodepool have a mature wrapper around zk, changing the backend in the context of this spec doesn't seem like a reasonable thing to do. If using zk become an issue, then I think it will be possible to adapt the wrapper for another backend. It seems better to wait and see if this spec will push zk to its limit before evaluating another backend like etcd.

However, it's unclear to me how the roadmap is going to be implemented, and if using zk to store the scheduler objects is going to happen progressively and/or if it could be controlled with an optional toggle that would be disabled by default until the feature is complete. Perhaps this should be mentioned in the spec?

Patch-set: 18
Reviewer: Gerrit User 9311 <9311@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2