Consolidate scheduler pause/exit as hibernation
The two properties, `_pause` and `_exit`, in scheduler are
both used for "hibernation", i.e., saving the queue in a
pickled file. The `resume` method is used to wake (load
the queue from the picked file) up the scheduler.

This is a preparation for pause/resume of a scheduler
which is needed in a multi-scheduler operation. This
hibernation functionality will be removed in the near
future and replaced with keeping the queue in Zookeeper.

Change-Id: I93343272b04eedcf10e963b3ba47042a287b6e9e
Story: 2007192