Update patch set 23
Patch Set 23:

To estimate the time/sizing requirements of this, I tested it on the openstack tenant using the repl.  It took 19.66 seconds to generate the dictionary data structure.  It has mappingproxy objects in it, so is not yet JSON serializable, but using repl to get an approximation returned a string 137MB in length.

I think we should revisit this after the HA scheduler work -- I'm still not sure we want to send 137MB responses to API clients, but at least if we do decide to do so, it won't affect the scheduler.

Patch-set: 23
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0