Evaluate CODEOWNERS settings during canMerge check
GitHub has a branch protection setting that allows to block PRs from
being merged until reviews of users that are responsible for the files
in the PR have reviewed. Since GitHub doesn't expose an API that tells
us whether we can merge a certain PR, we have to determine code ownership
ourselves. Therefore we download the CODEOWNERS files from
all possible locations within the repository and cache the parsed file.
When determining if the commit can be merged, we go through the list of
files that are part of the PR and determine which users or teams have to
provide an approval. We then look through the available review to check
if all the file groups have at least one review.

Please note that there is a limit of 300 files per PR. If a PR has more
than 300 files, we cannot determine whether the PR would be mergeable
and we fall back to the old behavior. This might therefore still lead to
merge failures on GitHub side.

Change-Id: I38f075eef8e89d341a0deed862dba32aac12801f
Co-Authored-By: Jan Kubovy <jan.kubovy@bmw.de>
Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>