Update patch set 3
Patch Set 3:

Thanks for the response, Matthieu. I will have a look. The main reason for this is that we could use it to dequeue buildsets via the Github checks API later on. Github supports checks to have an "external_id" which is the identifier of this check in the integrator's system (zuul in our case). Using the buildset ID as external_id is a perfect match. Thus, using this directly to dequeue a change is the simplest approach rather than looking up all necessary parameters like tenant, pipeline, project, ... (which we might not even have on the Github side, because Github doesn't need to care about those).

Patch-set: 3
Label: Verified=0