Update patch set 18
Patch Set 18:

Currently, on start or full-reconfigure, Zuul reads configuration from every branch of every project it's been told to look at. When it receives an event from the code review system, it attempts to match the details of the event against that combined body of configuration to decide what jobs should be run. It is not aware of selection criteria within the source code represented by a tag, especially one for which it's only just seen a push event, as the events themselves consist only of metadata.

Patch-set: 18