Update patch set 4
Patch Set 4:

I accidentally reused the change-id of the former change and overwrote it with a completely different commit. Originally, the previous version of this change was meant as a requirement for this new commit, but since we talked the topic over again, it's not needed anymore. Thus, this is now the new change which follows up on the github checks api

Patch-set: 4
Label: Verified=0