Allow check runs to be configured as required status in pipeline config
Although commit checks and statuses can only be retrieved via their
respective APIs, Github does not differentiate between both in terms of
branch protection and in the status section (below the comments of a
PR).

To mimic this behaviour in Zuul, one can now configure a check run as
required status in the pipeline config.

Change-Id: Ia5757c476bcee6082991928ab7c1743d0200d04a