Update patch set 18
Patch Set 18:

> > Though thinking about it a little more, what you're describing is
 > > similar to the temporary speculative layouts Zuul compiles for
 > > pre-merge pipelines.
 > 
 > Yes, that is what I ment. (Sorry for being unclear.)
 > I'd guess it would be decently easy too: return the original commit
 > without any merging.
 > Simple and unambiguous.

It's not that simple because jobs can span over many repos where git tags cannot be pushed atomically. I agree that this is not a 100% solution yet, but the best we can get without a large effort. So I'd vote for merging this for now and think after that how we can cover the other use cases.

Patch-set: 18