Declare support for Python3.8
Replace our tox-py36 job with tox-py38, extend the list of trove
classifiers for Python versions in package metadata, and replace the
"py35" in the tox.ini envlist with just "py3" so that folks running
`tox` unqualified on their systems will use whatever python3
interpreter they have on hand (odds are it's in our supported range
these days). Also uncap python-daemon so we use a version compatible
with Python >=3.8.

Co-Authored-By: Jeremy Stanley <fungi@yuggoth.org>

Change-Id: Id45c799ec6b774056d8977c13f59bc98543ef5f5